+++ poetry/version/version_selector.py
@@ -26,8 +26,9 @@ class VersionSelector(object):
             constraint = parse_constraint("*")
 
         candidates = self._pool.find_packages(
-            package_name, constraint, allow_prereleases=allow_prereleases
+            package_name, constraint, allow_prereleases=True
         )
+        only_prereleases = all([c.version.is_prerelease() for c in candidates])
 
         if not candidates:
             return False
@@ -37,7 +38,12 @@ class VersionSelector(object):
         # Select highest version if we have many
         package = candidates[0]
         for candidate in candidates:
-            if candidate.is_prerelease() and not dependency.allows_prereleases():
+            if (
+                candidate.is_prerelease()
+                and not dependency.allows_prereleases()
+                and not allow_prereleases
+                and not only_prereleases
+            ):
                 continue
 
             # Select highest version of the two
@@ -52,24 +58,20 @@ class VersionSelector(object):
         return self._transform_version(version.text, package.pretty_version)
 
     def _transform_version(self, version, pretty_version):
-        # attempt to transform 2.1.1 to 2.1
-        # this allows you to upgrade through minor versions
         try:
             parsed = Version.parse(version)
             parts = [parsed.major, parsed.minor, parsed.patch]
         except ValueError:
             return pretty_version
 
-        # check to see if we have a semver-looking version
-        if len(parts) == 3:
-            # remove the last parts (the patch version number and any extra)
-            if parts[0] != 0:
-                del parts[2]
+        parts = parts[: parsed.precision]
 
+        # check to see if we have a semver-looking version
+        if len(parts) < 3:
+            version = pretty_version
+        else:
             version = ".".join(str(p) for p in parts)
             if parsed.is_prerelease():
                 version += "-{}".format(".".join(str(p) for p in parsed.prerelease))
-        else:
-            return pretty_version
 
         return "^{}".format(version)